feat(machine): instructions are aligned to 64 bits

+++ opcode/index.go
@@ -1,7 +1,7 @@
 package opcode
 
 const (
-	NOP uint8 = iota // DO NOTHING
+	NOP uint32 = iota // DO NOTHING
 
 	/* Stack manipulation */
